+++ www/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport"
           content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>httprouter-example</title>
+    <title>httprouter-util</title>
 
     <link rel="shortcut icon" type="image/png" href="favicon.ico"/>
 </head>
